+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2019-09-27
+### Added
+- 一鍵切換 All
+- Get Last Scan Log
+
+### Changed
+- Scan Data mapping funciton change to Service
+- Delete CDN Provider Function
+- Check Default CDN output add param
+
+### Removed
+- 根據 Region 選擇 A 切換至 B CDN Provider
+
+### Fixed
+- Domain Cname 轉小寫英文
+
 ## [0.3.3] - 2019-09-11
 ### Fixed
 - Domain added to group Bug Fix